Change implementation of assignments
ShiftRoleAssignment and WorkerRoleAssignment create multiple Shift and
Worker objects of the same shift and worker.

The circular dependency makes it hard to create a TypicalShift
and TypicalWorker to test the functionality.

Let's remove the circular dependency and implement an Assignment class
that is dependent on the Shift, Worker and Role classes.